Move Clipboard to ui/
Belong here as it interacts with DOM

Part of #1507
