+++ src/ui/Clipboard.test.ts
